pyproject.toml
[project]
name = "sigrene"
version = "0.1.0"
description = ""
authors = [
    {name = "DavidMunozValero",email = "david.munoz6@alu.uclm.es"}
]
license = {text = "UNLICENSED"}
requires-python = ">=3.12"
dependencies = [
    "pymongo (>=4.16.0,<5.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "fastapi[standard] (>=0.135.1,<0.136.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "bcrypt (==3.2.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
